Updating convert script.
It looks like in early 2020 the
nbconvert system has been rewritten
or at least substantially changed.

The method outlined in the readme
for converting to gwoptics scripts
scripts/HTML_conversion/README
does not work any more and fails
claiming it cannot find basic.tpl

I think the 'new' approach is to
define a template folder with a
config file.

I have added this scripts/gwoptics
and it is based on 'classic' which
replicates the default jupyter notebook
look.

I could have based it on 'classic'
and imported. But has too much
flexibility and is probably
too complicated for what is needed here.