Fixed toc issues.
The script now uses another jinja
template to generate the extra pages.
This is very similar to the gwoptics
jinga template for converting from ipython
notebooks, but not exactly the same.

It would have been possible to do these fixes
without moving to jinja but since it is installed
I have kept to a common framework.